backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.3.0
numpy>=1.26.0